move getExplorerUrl to helpers/chain.ts
